Added Implementation Files, edited Test Cases
I added Implementation Files for the City, Unit, and Tile classes.

I edited several functions in the GameImpl classes so that the world would be properly set up according to the requirements for AlphaCiv.

These classes were set so that the test cases would all pass.

Edited a few test calls that didn't call the getString() method of each class.